backend/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.63.3
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-graphql-auth==0.3.16
django-graphql-jwt==0.4.0
django-oauth-toolkit==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
drf-social-oauth2==2.3.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
graphene==3.3
graphene-django==3.2.2
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
httplib2==0.22.0
idna==3.7
importlib_metadata==8.0.0
jwcrypto==1.5.6
Markdown==3.6
mutations==0.4.0
oauth2client==4.1.3
oauthlib==3.2.2
pillow==10.3.0
promise==2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
Rx==3.2.0
singledispatch==4.1.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
